requirements.txt
# Smart Contract Vulnerability Detection Benchmark
# Requirements

# Core
pydantic>=2.0
pyyaml>=6.0
click>=8.0

# Google Cloud / Vertex AI
google-cloud-aiplatform>=1.38
google-auth>=2.0

# Model API clients
openai>=1.0  # Used for DeepSeek via Vertex AI (OpenAI-compatible)

# Async support
aiohttp>=3.8
aiofiles>=23.0

# Data processing
numpy>=1.24

# Development (optional)
pytest>=7.0
pytest-asyncio>=0.21
black
ruff
